# ci(run): accelerate workflows