--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -1456,7 +1456,6 @@ jobs:
       - name: Install OpenSSL
         if: ${{ contains(matrix.os, 'windows') }}
         run: |
-          choco install openssl -y --no-progress
           Get-Command openssl
           openssl version
           echo "OPENSSL_LIB_DIR=C:/Program Files/OpenSSL/lib" >> $env:GITHUB_ENV